Refactor codes and add features for event management system
Enhanced the Event and User models and services, deleted the TestController and TicketPurchaseResponse classes, and implemented new files for Category management and ForgotPasswordService. The Dockerfile for MySQL and PhpMyAdmin initialization, UserService for managing user details, and added code for mail confirmation have also been included. Updated dependencies in the pom.xml with additional annotation and Thymeleaf libraries.